clean up top level framework modules, add module descriptions
