+++ examples/annotator_demo.py
@@ -3,10 +3,10 @@ import logging
 from langchain_community.tools.tavily_search import TavilySearchResults
 
 from tapeagents.demo import Demo
-from tapeagents.dialog_tape import DialogTape, DialogContext
+from tapeagents.dialog_tape import DialogContext, DialogTape
 from tapeagents.environment import ToolEnvironment
 from tapeagents.llms import LiteLLM
-from tapeagents.rendering import BasicRenderer
+from tapeagents.renderers.basic import BasicRenderer
 
 from .annotator import GroundednessAnnotator
 from .openai_function_calling import FunctionCallingAgent
