+++ tapeagents/llm_function.py
@@ -1,3 +1,7 @@
+"""
+Optimizable llm functions.
+"""
+
 from __future__ import annotations
 
 from typing import Any, Type
