+++ tapeagents/llms.py
@@ -1,3 +1,7 @@
+"""
+Wrapper for interacting with external and hosted large language models (LLMs).
+"""
+
 from __future__ import annotations
 
 import datetime
@@ -134,7 +138,7 @@ class LLM(BaseModel, ABC):
 
         Args:
             prompt (Prompt): The prompt object containing messages to send to the LLM.
-            **kwargs: Additional arguments to pass to the underlying LLM implementation.
+            **kwargs (dict, optional): Additional arguments to pass to the underlying LLM implementation.
 
         Returns:
             LLMStream: A stream of LLM events containing the model's response.
@@ -176,9 +180,6 @@ class LLM(BaseModel, ABC):
             prompt (Prompt): The prompt object containing the input messages for the LLM.
             message (LLMOutput): The output message generated by the LLM.
             cached (bool, optional): Indicates whether the output was retrieved from cache. Defaults to False.
-
-        Returns:
-            None
         """
         llm_call = LLMCall(
             timestamp=datetime.datetime.now().isoformat(),
@@ -258,9 +259,6 @@ class CachedLLM(LLM):
         and adds them to the cache using the prompt key as index. Each entry is converted
         to an LLMEvent model before caching.
 
-        Returns:
-            None
-
         Side Effects:
             - Updates the internal cache with log data
             - Logs the total number of reindexed entries at INFO level
@@ -299,7 +297,7 @@ class CachedLLM(LLM):
 
         Args:
             prompt (Prompt): The prompt object containing messages to send to the LLM.
-            **kwargs: Additional arguments to pass to the underlying LLM implementation.
+            **kwargs (dict, optional): Additional arguments to pass to the underlying LLM implementation.
 
         Returns:
             LLMStream: A stream of LLM events containing the model's response.
@@ -420,8 +418,8 @@ class LiteLLM(CachedLLM):
         logic for creating the training text.
 
         Args:
-            *args: Variable length argument list.
-            **kwargs: Arbitrary keyword arguments.
+            *args (list): Variable length argument list.
+            **kwargs (dict, optional): Arbitrary keyword arguments.
 
         Returns:
             TrainingText: The generated training text.
@@ -763,13 +761,13 @@ class ReplayLLM(LLM):
         Create a ReplayLLM instance from an existing LLM and a SQLite database file.
 
         Args:
-            cls: The class to instantiate.
+            cls (Type): The class to instantiate.
             llm (LLM): The original LLM instance.
             run_dir (str): The directory where the SQLite database file is located.
             prompts_file (str, optional): The name of the SQLite database file. Defaults to DB_DEFAULT_FILENAME.
 
         Returns:
-            ReplayLLM: An instance of ReplayLLM initialized with the LLM calls from the SQLite database.
+            (ReplayLLM): An instance of ReplayLLM initialized with the LLM calls from the SQLite database.
 
         Raises:
             AssertionError: If the SQLite database file does not exist at the specified path.
@@ -809,7 +807,7 @@ class ReplayLLM(LLM):
 
         Args:
             prompt (Prompt): The prompt object containing the messages to be processed.
-            **kwargs: Additional keyword arguments.
+            **kwargs (dict, optional): Additional keyword arguments.
 
         Returns:
             LLMStream: A stream of LLM events containing the generated output.
@@ -930,7 +928,7 @@ def trainable_llm_make_training_text(prompt: Prompt, output: LLMOutput, tokenize
     Args:
         prompt (Prompt): The input prompt containing conversation messages.
         output (LLMOutput): The model's output/response.
-        tokenizer: The tokenizer used to format the conversation.
+        tokenizer (PreTrainedTokenizer): The tokenizer used to format the conversation.
 
     Returns:
         TrainingText: A dataclass containing:
