+++ tapeagents/prompting.py
@@ -1,3 +1,7 @@
+"""
+Utilities for converting between tape steps and LLM messages.
+"""
+
 from tapeagents.agent import Agent
 from tapeagents.core import AgentStep, Call, Observation, Pass, Prompt, Respond, SetNextNode, Step, Tape
 from tapeagents.dialog_tape import SystemStep, ToolResult
