+++ tapeagents/renderers/basic.py
@@ -4,60 +4,9 @@ from typing import Any, Type
 import yaml
 from pydantic import BaseModel
 
-from tapeagents.agent import Agent
-from tapeagents.container_executor import CodeBlock
-from tapeagents.core import Action, Episode, Observation, Step, Tape, Thought
-from tapeagents.dialog_tape import (
-    AssistantStep,
-    DialogContext,
-    DialogTape,
-    SystemStep,
-    ToolCalls,
-    ToolResult,
-    UserStep,
-)
-from tapeagents.environment import CodeExecutionResult, ExecuteCode
-from tapeagents.observe import LLMCall, retrieve_tape_llm_calls
-from tapeagents.view import Call, Respond
-
-YELLOW = "#ffffba"
-LIGHT_YELLOW = "#ffffdb"
-SAND = "#e5e5a7"
-WHITE = "#ffffff"
-PURPLE = "#E6E6FA"
-RED = "#ff7b65"
-GREEN = "#6edb8f"
-BLUE = "#bae1ff"
-
-
-def render_dialog_plain_text(tape: DialogTape | None) -> str:
-    """
-    Renders a dialog tape into a plain text format.
-
-    Takes a DialogTape object containing conversation steps and formats them into human-readable text,
-    with each dialog step on a new line prefixed by the speaker/action type.
-
-    Args:
-        tape (Union[DialogTape, None]): A DialogTape object containing conversation steps, or None
-
-    Returns:
-        str: A string containing the formatted dialog, with each step on a new line.
-            Returns empty string if tape is None.
-    """
-    if tape is None:
-        return ""
-    lines = []
-    for step in tape:
-        if isinstance(step, UserStep):
-            lines.append(f"User: {step.content}")
-        elif isinstance(step, AssistantStep):
-            lines.append(f"Assistant: {step.content}")
-        elif isinstance(step, ToolCalls):
-            for tc in step.tool_calls:
-                lines.append(f"Tool calls: {tc.function}")
-        elif isinstance(step, ToolResult):
-            lines.append(f"Tool result: {step.content}")
-    return "\n".join(lines)
+from tapeagents.core import Episode, Observation, Step, Tape
+from tapeagents.dialog_tape import UserStep
+from tapeagents.observe import LLMCall
 
 
 class BasicRenderer:
@@ -163,7 +112,7 @@ class BasicRenderer:
         Args:
             step (Step): The step object containing the data to be rendered
             index (int): The index of the current step
-            **kwargs: Additional keyword arguments for rendering customization
+            **kwargs (dict, optional): Additional keyword arguments for rendering customization
 
         Returns:
             str: The rendered step as a formatted string in box format
@@ -297,7 +246,7 @@ class BasicRenderer:
 
         Args:
             llm_call (Union[LLMCall, None]): An LLM call object containing prompt and output information.
-            If None, returns an empty string.
+                If None, returns an empty string.
 
         Returns:
             str: HTML string representation of the LLM call. The HTML contains:
@@ -342,253 +291,3 @@ class BasicRenderer:
                 </details>
                 </div>"""
         return html
-
-
-class PrettyRenderer(BasicRenderer):
-    """Rendering enhancements for a handful of known steps."""
-
-    def __init__(self, show_metadata=False, **kwargs):
-        self.show_metadata = show_metadata
-        super().__init__(**kwargs)
-
-    @property
-    def style(self):
-        return super().style + (
-            "<style>"
-            ".observation { background-color: #baffc9;; }"
-            ".error_observation { background-color: #dd0000; }"
-            ".action { background-color: #cccccc; }"
-            ".thought { background-color: #ffffdb; }"
-            ".call { background-color: #ffffff; }"
-            ".respond { background-color: #ffffff; }"
-            ".step-header { margin: 2pt 2pt 2pt 0 !important; }"
-            ".step-text { font-size: 12px; white-space: pre-wrap; word-wrap: break-word;}"
-            "</style>"
-        )
-
-    def render_step(self, step: Step, index: int, **kwargs):
-        title = type(step).__name__
-        if isinstance(step, UserStep):
-            role = "User"
-            title = ""
-            class_ = "observation"
-        elif isinstance(step, SystemStep):
-            role = "System"
-            title = ""
-            class_ = "observation"
-        elif isinstance(step, AssistantStep):
-            role = "Assistant"
-            title = ""
-            class_ = "action"
-        elif isinstance(step, DialogContext):
-            role = ""
-            class_ = "observation"
-        elif isinstance(step, Call):
-            role = ""
-            title = f"{step.metadata.agent.split('/')[-1]} calls {step.agent_name}"
-            class_ = "call"
-        elif isinstance(step, Respond):
-            role = ""
-            parts = step.metadata.agent.split("/")
-            title = f"{parts[-1]} responds to {parts[-2]}" if len(parts) > 1 else f"{step.metadata.agent} responds"
-            class_ = "respond"
-        elif isinstance(step, Thought):
-            role = "Thought"
-            class_ = "thought"
-        elif isinstance(step, Action):
-            role = "Action"
-            class_ = "action"
-        elif isinstance(step, CodeExecutionResult):
-            role = "Observation"
-            class_ = "error_observation" if step.result.exit_code != 0 else "observation"
-        elif isinstance(step, Observation):
-            role = "Observation"
-            class_ = "observation"
-        else:
-            raise ValueError(f"Unknown object type: {type(step)}")
-
-        dump = step.model_dump()
-        if not self.show_metadata:
-            dump.pop("metadata", None)
-
-        def pretty_yaml(d: dict):
-            return yaml.dump(d, sort_keys=False, indent=2) if d else ""
-
-        def maybe_fold(content: Any):
-            content = str(content)
-            summary = f"{len(content)} characters ..."
-            if len(content) > 1000:
-                return f"<details><summary>{summary}</summary>{content}</details>"
-            return content
-
-        if (content := getattr(step, "content", None)) is not None:
-            # TODO: also show metadata here
-            del dump["content"]
-            text = pretty_yaml(dump) + ("\n" + maybe_fold(content) if content else "")
-        elif isinstance(step, ExecuteCode):
-            del dump["code"]
-
-            def format_code_block(block: CodeBlock) -> str:
-                return f"```{block.language}\n{block.code}\n```"
-
-            code_blocks = "\n".join([format_code_block(block) for block in step.code])
-            text = pretty_yaml(dump) + "\n" + maybe_fold(code_blocks)
-        elif isinstance(step, CodeExecutionResult):
-            del dump["result"]["output"]
-            text = pretty_yaml(dump) + "\n" + maybe_fold(step.result.output)
-        else:
-            text = pretty_yaml(dump)
-
-        index_str = f"[{index}]"
-        header_text = title if not role else (role if not title else f"{role}: {title}")
-        header = f"{index_str} {header_text}"
-
-        return (
-            f"<div class='basic-renderer-box {class_}'>"
-            f"<h4 class='step-header'>{header}</h4>"
-            f"<pre class='step-text'>{text}</pre>"
-            f"</div>"
-        )
-
-
-def step_view(step: Step, trim: bool = False) -> str:
-    title = get_step_title(step)
-    text = get_step_text(step, trim)
-    return f"{title}:\n{text}"
-
-
-def get_step_title(step: Step | dict) -> str:
-    title = ""
-    step_dict = step if isinstance(step, dict) else step.model_dump()
-    if kind := step_dict.get("kind", None):
-        if kind.endswith("_thought"):
-            kind = kind[:-8]
-        elif kind.endswith("_action"):
-            kind = kind[:-7]
-        elif kind.endswith("_observation"):
-            kind = kind[:-12]
-        title = kind.replace("_", " ").title()
-    return title
-
-
-def get_step_text(step: Step | dict, trim: bool = False, exclude_fields={"kind", "role"}) -> str:
-    step_dict = step if isinstance(step, dict) else step.model_dump()
-    clean_dict = {
-        k: (v[:300] if trim and isinstance(v, str) else v)
-        for k, v in step_dict.items()
-        if k not in exclude_fields and v != [] and v != {}
-    }
-    if not clean_dict:
-        return ""
-    return to_pretty_str(clean_dict).strip()
-
-
-def to_pretty_str(a: Any, prefix: str = "", indent: int = 2) -> str:
-    """Convert any Python object to a pretty formatted string representation.
-
-    This function recursively formats nested data structures (lists and dictionaries)
-    with proper indentation and line breaks for improved readability.
-
-    Args:
-        a (Any): The object to be formatted. Can be a list, dictionary, or any other type.
-        prefix (str, optional): String to prepend to each line. Used for recursive indentation. Defaults to "".
-        indent (int, optional): Number of spaces to use for each level of indentation. Defaults to 2.
-
-    Returns:
-        str: A formatted string representation of the input object.
-
-    Example:
-        ```python
-        data = {"foo": [1, 2, {"bar": "baz"}]}
-        print(to_pretty_str(data))
-        ```
-        foo:
-          - 1
-          - 2
-          - bar: baz
-    """
-    view = ""
-    if isinstance(a, list) and len(a):
-        if len(str(a)) < 80:
-            view = str(a)
-        else:
-            lines = []
-            for item in a:
-                value_view = to_pretty_str(item, prefix + " " * indent)
-                if "\n" in value_view:
-                    value_view = f"\n{value_view}"
-                lines.append(f"{prefix}- " + value_view)
-            view = "\n".join(lines)
-    elif isinstance(a, dict) and len(a):
-        lines = []
-        for k, v in a.items():
-            value_view = to_pretty_str(v, prefix + " " * indent)
-            if "\n" in value_view:
-                value_view = f"\n{value_view}"
-            lines.append(f"{prefix}{k}: {value_view}")
-        view = "\n".join(lines)
-    else:
-        view = str(a)
-    return view
-
-
-def render_agent_tree(agent: Agent, show_nodes: bool = True, indent_increment: int = 4) -> str:
-    """
-    Renders an ASCII tree representation of an agent's hierarchical structure.
-
-    This function creates a visual tree diagram showing the relationships between agents,
-    their nodes, and subagents using ASCII characters. Each level is indented to show
-    the hierarchy.
-
-    Args:
-        agent (Agent): The root agent object to render the tree from.
-        show_nodes (bool, optional): Whether to display the agent's nodes in the tree.
-            Defaults to True.
-        indent_increment (int, optional): Number of spaces to indent each level.
-            Defaults to 4.
-
-    Returns:
-        str: A string containing the ASCII tree representation.
-
-    Example:
-        ```
-        > The Manager
-            .node1
-            .node2
-            > His Assistant 1
-                .node1
-                .node2
-            > His Helper 2
-                .node1
-                .node2
-        ```
-    """
-
-    def render(agent: Agent, indent: int = 0) -> str:
-        lines = [f"{' ' * indent}> {agent.name}"]
-        if show_nodes:
-            lines.extend([f"{' ' * (indent + indent_increment)}.{node.name}" for node in agent.nodes])
-        for subagent in agent.subagents:
-            lines.append(render(subagent, indent + indent_increment))
-        return "\n".join(lines)
-
-    return render(agent)
-
-
-def render_tape_with_prompts(tape: Tape, renderer: BasicRenderer):
-    """
-    Renders a tape with prompts using the specified renderer.
-
-    This function combines the tape's LLM calls with the renderer's style and rendering
-    to produce a complete rendered output of the tape's content.
-
-    Args:
-        tape (Tape): The tape object to be rendered
-        renderer (BasicRenderer): The renderer to use for rendering the tape
-
-    Returns:
-        str: The rendered tape content with applied styling, combining both
-             the renderer's style and the tape's content with LLM calls
-    """
-    llm_calls = retrieve_tape_llm_calls(tape)
-    return renderer.style + renderer.render_tape(tape, llm_calls)
