+++ tapeagents/tools/__init__.py
@@ -1,0 +1,3 @@
+"""
+Tools for the agents to use.
+"""
